requirements.txt
aioredis==1.3.1
async-timeout==3.0.1
asyncpg==0.20.1
atomicwrites==1.3.0
attrs==19.3.0
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
ConfigArgParse==1.1
databases==0.2.6
fastapi==0.49.0
Flask==1.1.1
gevent==1.5a3
geventhttpclient-wheels==1.3.1.dev2
greenlet==0.4.15
h11==0.9.0
hiredis==1.0.1
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
locustio==0.14.5
MarkupSafe==1.1.1
more-itertools==8.2.0
msgpack==1.0.0
packaging==20.3
passlib==1.7.2
pluggy==0.13.1
psutil==5.7.0
psycopg2==2.8.4
py==1.8.1
pycparser==2.19
pydantic==1.4
PyJWT==1.7.1
pyparsing==2.4.6
pytest==5.4.1
python-multipart==0.0.5
pyzmq==19.0.0
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.13
starlette==0.12.9
urllib3==1.25.8
uvicorn==0.11.3
wcwidth==0.1.9
websockets==8.1
Werkzeug==1.0.0
zipp==3.1.0
